Add TransactionAuth class with verifyTransactionOwnership() to transactionController methods

+++ src/controller/transactionController.js
@@ -13,8 +13,13 @@ export class TransactionController {
 
     static async get(req, res) {
         const { id } = req.params
+        const { id: USER_ID } = req.user
+
         try {
+            await TransactionAuth.verifyTransactionOwnership(id, USER_ID)
+
             const transaction = await Transaction.getTransactionById(id)
+            
             if (transaction) {
                 res.json(transaction)
             } else {
@@ -28,16 +33,19 @@ export class TransactionController {
     static async update(req, res) {
         const { id } = req.params
         const { type_transaction, amount, date_transaction, description_transaction } = req.body
-    
-        // Crear un objeto con los campos opcionales que se quieren actualizar
+        const { id: USER_ID } = req.user
         const data = {}
-        if (type_transaction) data.type_transaction = type_transaction
-        if (amount) data.amount = amount
-        if (date_transaction) data.date_transaction = date_transaction
-        if(description_transaction) data.description_transaction = description_transaction
-    
+
         try {
+            await TransactionAuth.verifyTransactionOwnership(id, USER_ID)
+            
+            if (type_transaction) data.type_transaction = type_transaction
+            if (amount) data.amount = amount
+            if (date_transaction) data.date_transaction = date_transaction
+            if (description_transaction) data.description_transaction = description_transaction
+            
             await Transaction.updateTransaction(id, data)
+            
             res.json({ message: 'Transaction updated successfully' })
         } catch (error) {
             res.status(500).json({ error: 'Error updating transaction' })
@@ -46,16 +54,19 @@ export class TransactionController {
 
     static async delete(req, res) {
         const { id } = req.params
+        const { id: USER_ID } = req.user
+
         try {
+            await TransactionAuth.verifyTransactionOwnership(id, USER_ID)
+
             await Transaction.deleteTransaction(id)
+
             res.json({ message: 'Transaction deleted successfully' })
         } catch (error) {
             res.status(500).json({ error: 'Error deleting transaction' })
         }
     }
 
-
-    // Obtener resumen del presupuesto
     static async getBudgetSummary(req, res) {
         const { id_user } = req.params // Se asume que el id_user se pasa como parámetro
         try {
@@ -64,5 +75,15 @@ export class TransactionController {
         } catch (error) {
             res.status(500).json({ error: 'Error fetching budget summary', error })
         }
+    }
+}
+
+export class TransactionAuth {
+    static async verifyTransactionOwnership(transaction_id, USER_ID) {
+        const TARGETED_TRANSACTION = await Task.getTransactionById(transaction_id)
+            
+        if (!TARGETED_TRANSACTION) throw new Error('Transaction not found')
+
+        if (TARGETED_TASK.id_user !== USER_ID) throw new Error('Invalid transaction n ID')
     }
 }
